refactor(#3479): extract recovered-turn analytics+transcript cluster from recovery_engine
Behavior-preserving move of 5 helpers (extract_turn_analytics_from_output,
recovered_turn_duration_ms, lookup_turn_finished_dispatch_kind,
recovered_transcript_turn_id, persist_recovered_transcript) into a new leaf
submodule recovery_engine/analytics_transcript.rs (verbatim bodies; only
pub(super) visibility + super::super:: path depth adjusted).

Root keeps byte-identical call sites via re-import. recovered_transcript_turn_id
is cfg(test)-gated since the root reaches it only from its unit test (production
calls it internally inside analytics_transcript) — avoids an unused-import
warning in non-test builds (codex r1).

recovery_engine.rs 3424 -> 3330 prod LoC; giant baseline + change-surfaces.md
freeze synced to 3330.
